requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
acres==0.2.0
    # via nipype
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
attrs==24.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.1
    # via streamlit
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
ci-info==0.3.0
    # via etelemetry
click==8.1.8
    # via
    #   nipype
    #   streamlit
configobj==5.0.9
    # via fitz
configparser==7.1.0
    # via fitz
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
etelemetry==0.3.1
    # via nipype
faiss-cpu==1.9.0.post1
    # via union-rep-assistant (pyproject.toml)
filelock==3.17.0
    # via nipype
fitz==0.0.1.dev2
    # via union-rep-assistant (pyproject.toml)
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via fitz
httpx==0.28.1
    # via
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isodate==0.6.1
    # via rdflib
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jiter==0.8.2
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
langchain==0.3.15
    # via langchain-community
langchain-community==0.3.15
    # via union-rep-assistant (pyproject.toml)
langchain-core==0.3.31
    # via
    #   langchain
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
langchain-openai==0.3.1
    # via union-rep-assistant (pyproject.toml)
langchain-text-splitters==0.3.5
    # via langchain
langsmith==0.3.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
looseversion==1.3.0
    # via nipype
lxml==5.3.0
    # via
    #   prov
    #   pyxnat
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.25.1
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
narwhals==1.23.0
    # via altair
networkx==3.4.2
    # via
    #   nipype
    #   prov
nibabel==5.3.2
    # via
    #   fitz
    #   nipype
nipype==1.9.2
    # via fitz
numpy==2.2.2
    # via
    #   faiss-cpu
    #   fitz
    #   langchain
    #   langchain-community
    #   nibabel
    #   nipype
    #   pandas
    #   pydeck
    #   scipy
    #   streamlit
openai==1.60.0
    # via langchain-openai
orjson==3.10.15
    # via langsmith
packaging==24.2
    # via
    #   altair
    #   etelemetry
    #   faiss-cpu
    #   langchain-core
    #   marshmallow
    #   nibabel
    #   nipype
    #   streamlit
pandas==2.2.3
    # via
    #   fitz
    #   streamlit
pathlib==1.0.1
    # via pyxnat
pillow==11.1.0
    # via streamlit
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
protobuf==5.29.3
    # via streamlit
prov==2.0.1
    # via nipype
puremagic==1.28
    # via nipype
pyarrow==19.0.0
    # via streamlit
pydantic==2.10.5
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.7.1
    # via langchain-community
pydeck==0.9.1
    # via streamlit
pydot==3.0.4
    # via nipype
pygments==2.19.1
    # via rich
pyparsing==3.2.1
    # via
    #   httplib2
    #   pydot
    #   rdflib
python-dateutil==2.9.0.post0
    # via
    #   nipype
    #   pandas
    #   prov
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.2
    # via pandas
pyxnat==1.6.2
    # via fitz
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
rdflib==6.3.2
    # via
    #   nipype
    #   prov
referencing==0.36.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   union-rep-assistant (pyproject.toml)
    #   etelemetry
    #   langchain
    #   langchain-community
    #   langsmith
    #   pyxnat
    #   requests-toolbelt
    #   streamlit
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
scipy==1.15.1
    # via
    #   fitz
    #   nipype
simplejson==3.19.3
    # via nipype
six==1.17.0
    # via
    #   isodate
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.37
    # via
    #   langchain
    #   langchain-community
streamlit==1.41.1
    # via union-rep-assistant (pyproject.toml)
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   streamlit
tiktoken==0.8.0
    # via langchain-openai
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
tqdm==4.67.1
    # via openai
traits==6.4.3
    # via nipype
typing-extensions==4.12.2
    # via
    #   altair
    #   anyio
    #   langchain-core
    #   nibabel
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   streamlit
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
watchdog==6.0.0
    # via streamlit
yarl==1.18.3
    # via aiohttp
zstandard==0.23.0
    # via langsmith
